Refactor localization configurations & models
