--- src/Modules/SimplCommerce.Module.Localization/ResourceString.cs
@@ -1,11 +1,0 @@
-﻿namespace SimplCommerce.Module.Localization
-{
-    public class ResourceString
-    {
-        public string Culture { get; set; }
-
-        public string Key { get; set; }
-
-        public string Value { get; set; }
-    }
-}
