+++ src/SimplCommerce.Infrastructure/GlobalConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using SimplCommerce.Infrastructure.Localization;
 
 namespace SimplCommerce.Infrastructure
 {
@@ -11,7 +12,7 @@ namespace SimplCommerce.Infrastructure
 
         public static IList<ModuleInfo> Modules { get; set; } = new List<ModuleInfo>();
 
-        public static IList<SimpleCulture> SimpleCultures { get; set; } = new List<SimpleCulture>();
+        public static IList<Culture> Cultures { get; set; } = new List<Culture>();
 
         public static string WebRootPath { get; set; }
