+++ src/SimplCommerce.Infrastructure/Localization/Culture.cs
@@ -1,9 +1,11 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using SimplCommerce.Infrastructure.Models;
 
-namespace SimplCommerce.Module.Localization.Models
+namespace SimplCommerce.Infrastructure.Localization
 {
+    [Table("Localization_Culture")]
     public class Culture : EntityBaseWithTypedId<string>
     {
         public Culture(string id)
