--- src/SimplCommerce.Infrastructure/SimpleCulture.cs
@@ -1,9 +1,0 @@
-﻿namespace SimplCommerce.Infrastructure
-{
-    public class SimpleCulture
-    {
-        public string Id { get; set; }
-
-        public string Name { get; set; }
-    }
-}
