+++ src/SimplCommerce.WebHost/Extensions/ApplicationBuilderExtensions.cs
@@ -12,8 +12,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Net.Http.Headers;
 using SimplCommerce.Module.Core.Extensions;
 using SimplCommerce.Infrastructure.Data;
-using SimplCommerce.Module.Localization.Models;
 using SimplCommerce.Infrastructure;
+using SimplCommerce.Infrastructure.Localization;
 
 namespace SimplCommerce.WebHost.Extensions
 {
@@ -101,22 +101,10 @@ namespace SimplCommerce.WebHost.Extensions
         public static IApplicationBuilder UseCustomizedRequestLocalization(this IApplicationBuilder app)
         {
             var cultureRepository = app.ApplicationServices.GetRequiredService<IRepositoryWithTypedId<Culture, string>>();
-            var cultures = cultureRepository.Query().ToList();
+            GlobalConfiguration.Cultures = cultureRepository.Query().ToList();
+            app.UseRequestLocalization(GlobalConfiguration.Cultures
+                .OrderBy(c => c.IsDefault).Select(c => c.Id).ToArray());
 
-            GlobalConfiguration.SimpleCultures = cultures.Select(x => new SimpleCulture { Id = x.Id, Name = x.Name }).ToList();
-            var supportedCultures = cultures.Select(x => new CultureInfo(x.Id)).ToList();
-            var defaultCulture = cultures.Where(x => x.IsDefault).Select(x => new CultureInfo(x.Id)).FirstOrDefault();
-            if(defaultCulture == null)
-            {
-                defaultCulture = new CultureInfo("en-US");
-            }
-
-            app.UseRequestLocalization(new RequestLocalizationOptions
-            {
-                DefaultRequestCulture = new RequestCulture(defaultCulture, defaultCulture),
-                SupportedCultures = supportedCultures,
-                SupportedUICultures = supportedCultures
-            });
             return app;
         }
     }
